requirements.txt
google-api-python-client==2.149.0
gspread==6.1.4
google-auth==2.36.0
oauth2client==4.1.3
youtube-transcript-api==0.6.2
